backend/requirements.txt
# App
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic-settings==2.5.2
di4injector==0.0.2      # DI
slf4py==0.0.4           # Logger
html2text==2024.2.26    # Mail
requests==2.32.3        # HTTP Request
python-dotenv==1.0.1    # Environment
pytz==2024.2            # Timezone
cachetools==5.5.0       # Cache
SQLAlchemy==2.0.36      # ORM
#mysqlclient==2.2.4      # MySQL
psycopg2-binary==2.9.9  # Postgresql
psycopg2==2.9.9         # Postgresql
redis[hiredis]          # Redis

# API Gateway
PyJWT==2.9.0  # JWT

# Authority
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0